* GormCore/GormClassInspector.m 	* GormCore/GormClassManager.m 	* GormCore/GormDocument.m 	* GormCore/GormGenericEditor.m 	* GormCore/GormObjectInspector.m 	* GormCore/GormOutlineView.m 	* GormCore/GormPrivate.h 	* GormCore/GormPrivate.m 	* Palettes/0Menus/GormMenuEditor.m 	* Palettes/1Windows/GormDrawerAttributesInspector.m 	* Palettes/3Containers/GormTableColumnAttributesInspector.m 	* Palettes/3Containers/GormTableViewEditor.m 	* Palettes/4Data/DataPalette.m 	* Palettes/4Data/GormNumberFormatterAttributesInspector.m: Change int/unsigned int to NSInteger 	and NSUInteger to address 64-bit issues.  Patch by 	Sebastian Reitenbach <sebastia@l00-bugdead-prods.de> 	* Plugins/Gorm/GormGormWrapperLoader.m 	* Plugins/Nib/GormNibWrapperBuilder.m 	* Plugins/Nib/GormNibWrapperLoader.m 	* Plugins/Xib/GormXibWrapperLoader.m: Clean up warnings found by clang.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@35091 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassManager.m
@@ -1117,7 +1117,7 @@
 
   if (classInfo != nil && [classInformation objectForKey: name] == nil)
     {
-      int index = 0;
+      NSUInteger index = 0;
       NSArray *subclasses = [self subClassesOf: oldName];
 
       RETAIN(classInfo); // prevent loss of the information...
